ngeo-api/api/ngeo/apps/common/serializers.py
from rest_framework import serializers
from notifications.models import Notification
from .models import Area


class AreaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Area
        fields = "__all__"


class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = "__all__"
        depth = 2
