ngeo-api/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API for NGEO application"
authors = ["Omulosi <mulongojohnpaul@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
amqp = "5.0.5"
appdirs = "1.4.4"
asgiref = "3.3.1"
astroid = "2.4.2"
attrs = "20.3.0"
backcall = "0.2.0"
bandit = "1.7.0"
billiard = "3.6.3.0"
black = "20.8b1"
celery = "5.0.5"
certifi = "2020.12.5"
cffi = "1.14.5"
chardet = "4.0.0"
click = "7.1.2"
click-didyoumean = "0.0.3"
click-plugins = "1.1.1"
click-repl = "0.1.6"
coreapi = "2.3.3"
coreschema = "0.0.4"
coverage = "5.4"
cryptography = "3.4.6"
decorator = "4.4.2"
defusedxml = "0.7.1"
dj-database-url = "0.5.0"
dj-rest-auth = "2.1.3"
Django = "3.1.5"
django-cors-headers = "3.7.0"
django-createsuperuser = "2020.12.3"
django-debug-toolbar = "3.2"
django-environ = "0.4.5"
django-expiring-token = "1.0.2"
django-extensions = "3.1.0"
django-filter = "2.4.0"
django-model-utils = "4.1.1"
django-notifications-hq = "1.6.0"
django-safedelete = "1.0.0"
django-templated-mail = "1.1.1"
django-workers = "0.2.0"
djangorestframework = "3.12.2"
djangorestframework-camel-case = "1.2.0"
djangorestframework-filters = "1.0.0.dev2"
djangorestframework-gis = "0.17"
djangorestframework-jsonapi = "4.0.0"
djangorestframework-simplejwt = "4.6.0"
djoser = "2.1.0"
drf-api-tracking = "1.7.8"
drf-pretty-exception-handler = "0.1.1"
et-xmlfile = "1.1.0"
gitdb = "4.0.5"
GitPython = "3.1.12"
gunicorn = "20.0.4"
idna = "2.10"
importlib-metadata = "1.7.0"
inflection = "0.5.1"
iniconfig = "1.1.1"
ipdb = "0.13.4"
ipython = "7.20.0"
ipython_genutils = "0.2.0"
isort = "5.7.0"
itypes = "1.2.0"
jedi = "0.18.0"
Jinja2 = "2.11.3"
jsonfield = "3.1.0"
kombu = "5.0.2"
lazy-object-proxy = "1.4.3"
leaflet = "0.0.3"
MarkupSafe = "1.1.1"
mccabe = "0.6.1"
mypy-extensions = "0.4.3"
numpy = "1.20.3"
oauthlib = "3.1.0"
openpyxl = "3.0.7"
packaging = "20.9"
pandas = "1.2.4"
parso = "0.8.1"
pathspec = "0.8.1"
pbr = "5.5.1"
pexpect = "4.8.0"
pickleshare = "0.7.5"
pip-tools = "5.5.0"
pluggy = "0.13.1"
prompt-toolkit = "3.0.14"
psycopg2-binary = "2.8.6"
ptyprocess = "0.7.0"
py = "1.10.0"
pycodestyle = "2.6.0"
pycparser = "2.20"
pydocstyle = "5.1.1"
pyflakes = "2.2.0"
Pygments = "2.7.4"
PyJWT = "2.0.1"
pylama = "7.7.1"
pylint = "2.6.0"
pyparsing = "2.4.7"
pytest = "6.2.2"
pytest-bandit = "0.5.2"
pytest-cov = "2.11.1"
pytest-django = "4.1.0"
python-dateutil = "2.8.1"
python3-openid = "3.2.0"
pytz = "2020.5"
PyYAML = "5.4.1"
raven = "6.10.0"
regex = "2020.11.13"
requests = "2.25.1"
requests-oauthlib = "1.3.0"
six = "1.15.0"
smmap = "3.0.5"
snowballstemmer = "2.1.0"
social-auth-app-django = "4.0.0"
social-auth-core = "4.1.0"
sqlparse = "0.4.1"
stevedore = "3.3.0"
swapper = "1.1.2.post1"
toml = "0.10.2"
traitlets = "5.0.5"
typed-ast = "1.4.2"
typing-extensions = "3.7.4.3"
uritemplate = "3.0.1"
urllib3 = "1.26.3"
vine = "5.0.0"
wcwidth = "0.2.5"
Werkzeug = "1.0.1"
whitenoise = "5.2.0"
wrapt = "1.12.1"
xlrd = "2.0.1"
zipp = "3.4.0"
pip = "^21.3.1"
django-axes = "^5.27.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
